customer rate
