+++ src/entities/customerEntity.py
@@ -1,6 +1,5 @@
 import json
 from collections import namedtuple
-from src.entities.customerAddressEntity import customerAddressEntity
  
 class customerEntity:
 
@@ -37,3 +36,40 @@ class customerEntity:
             _entity.main = us.main
             _customer_address.append(_entity)
         self.customer_address = _customer_address 
+
+class customerAddressEntity:
+
+    def __init__(self,id=0,id_customer=None,address=None,longitude=None,latitude=None,
+                main=None):
+        self.id = id
+        self.id_customer = id_customer
+        self.address = address
+        self.longitude = longitude
+        self.latitude = latitude
+        self.main = main
+
+    def toJSON(self):
+        return json.dumps(self, default=lambda o: o.__dict__,sort_keys=True, indent=4)
+        
+class customerRateEntity:
+
+    def __init__(self,id_user=0,id_service=None,id_customer=None,rate=None,description=None):
+        self.id_user = id_user
+        self.id_service = id_service
+        self.id_customer = id_customer
+        self.rate = rate
+        self.description = description
+        
+    def toJSON(self):
+        return json.dumps(self, default=lambda o: o.__dict__,sort_keys=True, indent=4)
+    
+    def requestToClass(self,resquest):
+        data = resquest.get_json() 
+        data = json.dumps(data)
+        values = json.loads(data, object_hook=lambda d: namedtuple('X', d.keys())(*d.values()))
+        self.id_user = values.id_user
+        self.id_service = values.id_service 
+        self.id_customer = values.id_customer
+        self.rate = values.rate
+        self.description = values.description
+
